+++ src/frontend/components/InstructorWrapper/index.spec.tsx
@@ -1,8 +1,6 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { Provider } from 'react-redux';
 
-import { bootstrapStore } from '../../data/bootstrapStore';
 import { appState } from '../../types/AppData';
 import { InstructorWrapper } from './index';
 
@@ -27,19 +25,12 @@ jest.mock('../../data/appData', () => ({
 }));
 
 describe('<InstructorWrapper />', () => {
-  const state = {
-    state: appState.INSTRUCTOR,
-    video: null,
-  } as any;
-
   it('wraps its children in an instructor view if the current user is an instructor', () => {
     mockState = appState.INSTRUCTOR;
     const { getByText, getByTitle } = render(
-      <Provider store={bootstrapStore(state)}>
-        <InstructorWrapper>
-          <div title="some-child" />
-        </InstructorWrapper>
-      </Provider>,
+      <InstructorWrapper>
+        <div title="some-child" />
+      </InstructorWrapper>,
     );
 
     getByText('InstructorView');
@@ -49,11 +40,9 @@ describe('<InstructorWrapper />', () => {
   it('just renders the children if the current user is not an instructor', () => {
     mockState = appState.STUDENT;
     const { getByTitle, queryByText } = render(
-      <Provider store={bootstrapStore(state)}>
-        <InstructorWrapper>
-          <div title="some-child" />
-        </InstructorWrapper>
-      </Provider>,
+      <InstructorWrapper>
+        <div title="some-child" />
+      </InstructorWrapper>,
     );
 
     expect(queryByText('InstructorView')).toBeNull();