+++ src/frontend/data/stores/useThumbnail/index.tsx
@@ -14,30 +14,32 @@ interface ThumbnailState extends StoreState<Thumbnail> {
   };
 }
 
-export const [useThumbnail, useThumbnailApi] = create<ThumbnailState>((set, get) => {
-  let thumbnails = {};
+export const [useThumbnail, useThumbnailApi] = create<ThumbnailState>(
+  (set, get) => {
+    let thumbnails = {};
 
-  if (appData.video && appData.video.thumbnail !== null) {
-    thumbnails = {
-      [appData.video.thumbnail.id]: appData.video.thumbnail,
+    if (appData.video && appData.video.thumbnail !== null) {
+      thumbnails = {
+        [appData.video.thumbnail.id]: appData.video.thumbnail,
+      };
+    }
+
+    return {
+      addMultipleResources: (thumbnailsToAdd: Thumbnail[]) =>
+        set(addMultipleResources(get(), modelName.THUMBNAIL, thumbnailsToAdd)),
+      addResource: (thumbnail: Thumbnail) =>
+        set(addResource<Thumbnail>(get(), modelName.THUMBNAIL, thumbnail)),
+      getThumbnail: () => {
+        if (get()[modelName.THUMBNAIL]) {
+          const thumbnailId = Object.keys(get()[modelName.THUMBNAIL]).shift();
+          return get()[modelName.THUMBNAIL][thumbnailId!];
+        }
+
+        return null;
+      },
+      removeResource: (thumbnail: Thumbnail) =>
+        set(removeResource(get(), modelName.THUMBNAIL, thumbnail)),
+      [modelName.THUMBNAIL]: thumbnails,
     };
-  }
-
-  return {
-    addMultipleResources: (thumbnailsToAdd: Thumbnail[]) =>
-      set(addMultipleResources(get(), modelName.THUMBNAIL, thumbnailsToAdd)),
-    addResource: (thumbnail: Thumbnail) =>
-      set(addResource<Thumbnail>(get(), modelName.THUMBNAIL, thumbnail)),
-    getThumbnail: () => {
-      if (get()[modelName.THUMBNAIL]) {
-        const thumbnailId = Object.keys(get()[modelName.THUMBNAIL]).shift();
-        return get()[modelName.THUMBNAIL][thumbnailId!];
-      }
-
-      return null;
-    },
-    removeResource: (thumbnail: Thumbnail) =>
-      set(removeResource(get(), modelName.THUMBNAIL, thumbnail)),
-    [modelName.THUMBNAIL]: thumbnails,
-  };
-});
+  },
+);