+++ src/frontend/data/stores/useVideo/index.spec.tsx
@@ -94,26 +94,38 @@ describe('stores/useVideo', () => {
   it('adds a resource to the store', () => {
     useVideoApi.getState().addResource({ id: 'newResource' } as any);
 
-    expect(useVideoApi.getState()[modelName.VIDEOS].newResource).toEqual({ id: 'newResource' });
+    expect(useVideoApi.getState()[modelName.VIDEOS].newResource).toEqual({
+      id: 'newResource',
+    });
   });
   it('removes an existing resource', () => {
     useVideoApi.getState().addResource({ id: 'toDelete' } as any);
 
-    expect(useVideoApi.getState()[modelName.VIDEOS].toDelete).toEqual({ id: 'toDelete' });
+    expect(useVideoApi.getState()[modelName.VIDEOS].toDelete).toEqual({
+      id: 'toDelete',
+    });
 
     useVideoApi.getState().removeResource({ id: 'toDelete' } as any);
 
     expect(useVideoApi.getState()[modelName.VIDEOS].toDelete).toBeUndefined();
   });
   it('adds multiple resources to the store', () => {
-    useVideoApi.getState().addMultipleResources([
-      { id: 'multi1' } as any,
-      { id: 'multi2' } as any,
-      { id: 'multi3' } as any,
-    ]);
+    useVideoApi
+      .getState()
+      .addMultipleResources([
+        { id: 'multi1' } as any,
+        { id: 'multi2' } as any,
+        { id: 'multi3' } as any,
+      ]);
 
-    expect(useVideoApi.getState()[modelName.VIDEOS].multi1).toEqual({ id: 'multi1' });
-    expect(useVideoApi.getState()[modelName.VIDEOS].multi2).toEqual({ id: 'multi2' });
-    expect(useVideoApi.getState()[modelName.VIDEOS].multi3).toEqual({ id: 'multi3' });
+    expect(useVideoApi.getState()[modelName.VIDEOS].multi1).toEqual({
+      id: 'multi1',
+    });
+    expect(useVideoApi.getState()[modelName.VIDEOS].multi2).toEqual({
+      id: 'multi2',
+    });
+    expect(useVideoApi.getState()[modelName.VIDEOS].multi3).toEqual({
+      id: 'multi3',
+    });
   });
 });